integrate non-working storage client into streamer
