+++ internal/bigquery/storage/managedwriter/stream_type.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package storage
+package managedwriter
 
 import (
 	storagepb "google.golang.org/genproto/googleapis/cloud/bigquery/storage/v1"
