+++ internal/bigquery/storage/managedwriter/writer_option.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package storage
+package managedwriter
 
 import "google.golang.org/protobuf/types/descriptorpb"
